src/pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "georisk"
version = "0.1.0"
description = "Geospatial risk analysis pipeline for land-surface change detection"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "GeoRisk Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pystac-client>=0.7.0",
    "planetary-computer>=1.0.0",
    "rasterio>=1.3.0",
    "rioxarray>=0.15.0",
    "xarray>=2024.1.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "shapely>=2.0.0",
    "geopandas>=0.14.0",
    "pyproj>=3.6.0",
    "boto3>=1.34.0",
    "httpx>=0.26.0",
    "click>=8.1.0",
    "structlog>=24.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
georisk = "georisk.cli:cli"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
